refactor: split mcp_server.py per-domain (#559)
Pure relocation of the 3,454-line mcp_server.py monolith (107 @mcp.tool()
functions across ~37 domains) into per-domain modules, mirroring the #495
CLI split.

- mcp_core.py holds the shared `mcp = FastMCP(...)` instance + helpers
  (run_agentwire_cmd, _mcp_result, _delivery_result, get_caller_session,
  get_portal_url, format_* formatters).
- 23 mcp_<domain>.py modules each `from .mcp_core import mcp`; domain-local
  helpers (format_session_contexts, _portal_request, _SAY_DESCRIPTION,
  _council_tag, _format_voice_status, TTS prompt) travel with their domain.
- mcp_server.py is now a thin 51-line module that imports every domain
  module so the @mcp.tool() decorators register on import, plus run_server().

No tool behavior, name, signature, or docstring changed — all 107 tools
still register (verified: mcp._tool_manager._tools == 107). Integration/unit
patch targets repointed to the symbols' new home modules (no assertion
changes). Full suite (2130 passed) + ruff clean.

Closes #559

Built by dotdev.dev
