+++ tests/unit/test_mcp_tools_args.py
@@ -17,9 +17,9 @@ def _failure(error="something broke"):
 
 
 class TestSessionTools:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_session.run_agentwire_cmd")
     def test_sessions_list_success(self, mock_cmd):
-        from agentwire.mcp_server import sessions_list
+        from agentwire.mcp_session import sessions_list
         mock_cmd.return_value = _success(sessions=[
             {"name": "app", "machine": "local", "windows": 1, "path": "/p", "type": "claude-bypass"},
         ])
@@ -27,33 +27,33 @@ class TestSessionTools:
         mock_cmd.assert_called_once_with(["list", "--sessions"])
         assert "app" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_session.run_agentwire_cmd")
     def test_sessions_list_failure(self, mock_cmd):
-        from agentwire.mcp_server import sessions_list
+        from agentwire.mcp_session import sessions_list
         mock_cmd.return_value = _failure("tmux not running")
         result = sessions_list()
         assert "Failed" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_session.run_agentwire_cmd")
     def test_session_create_minimal(self, mock_cmd):
-        from agentwire.mcp_server import session_create
+        from agentwire.mcp_session import session_create
         mock_cmd.return_value = _success()
         result = session_create(name="test")
         mock_cmd.assert_called_once_with(["new", "-s", "test"])
         assert "created" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_session.run_agentwire_cmd")
     def test_session_create_all_args(self, mock_cmd):
-        from agentwire.mcp_server import session_create
+        from agentwire.mcp_session import session_create
         mock_cmd.return_value = _success()
         session_create(name="x", project_dir="/p", roles="voice,worker", session_type="bare")
         args = mock_cmd.call_args[0][0]
         assert args == ["new", "-s", "x", "-p", "/p", "--roles", "voice,worker", "--type", "bare"]
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
-    @patch("agentwire.mcp_server.get_caller_session", return_value="orchestrator")
+    @patch("agentwire.mcp_session.run_agentwire_cmd")
+    @patch("agentwire.mcp_session.get_caller_session", return_value="orchestrator")
     def test_session_send_cross_session(self, mock_caller, mock_cmd):
-        from agentwire.mcp_server import session_send
+        from agentwire.mcp_session import session_send
         mock_cmd.return_value = _success(verified=True)
         session_send(session="worker", message="do task")
         args = mock_cmd.call_args[0][0]
@@ -62,10 +62,10 @@ class TestSessionTools:
         assert '[MESSAGE FROM SESSION "orchestrator"' in sent_msg
         assert 'session_send(session="orchestrator"' in sent_msg
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
-    @patch("agentwire.mcp_server.get_caller_session", return_value=None)
+    @patch("agentwire.mcp_session.run_agentwire_cmd")
+    @patch("agentwire.mcp_session.get_caller_session", return_value=None)
     def test_session_send_no_caller(self, mock_caller, mock_cmd):
-        from agentwire.mcp_server import session_send
+        from agentwire.mcp_session import session_send
         mock_cmd.return_value = _success(verified=True)
         result = session_send(session="target", message="hello")
         args = mock_cmd.call_args[0][0]
@@ -79,32 +79,32 @@ class TestSessionTools:
 
 
 class TestPaneTools:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_pane.run_agentwire_cmd")
     def test_pane_send_with_session(self, mock_cmd):
-        from agentwire.mcp_server import pane_send
+        from agentwire.mcp_pane import pane_send
         mock_cmd.return_value = _success(verified=True)
         pane_send(pane=1, message="task", session="my-session")
         args = mock_cmd.call_args[0][0]
         assert args == ["send", "--pane", "1", "--verify", "task", "-s", "my-session"]
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_pane.run_agentwire_cmd")
     def test_pane_send_without_session(self, mock_cmd):
-        from agentwire.mcp_server import pane_send
+        from agentwire.mcp_pane import pane_send
         mock_cmd.return_value = _success(verified=True)
         pane_send(pane=0, message="hi")
         args = mock_cmd.call_args[0][0]
         assert args == ["send", "--pane", "0", "--verify", "hi"]
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_pane.run_agentwire_cmd")
     def test_pane_output_success(self, mock_cmd):
-        from agentwire.mcp_server import pane_output
+        from agentwire.mcp_pane import pane_output
         mock_cmd.return_value = _success(output="some output")
         result = pane_output(pane=1)
         assert result == "some output"
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_pane.run_agentwire_cmd")
     def test_pane_output_failure(self, mock_cmd):
-        from agentwire.mcp_server import pane_output
+        from agentwire.mcp_pane import pane_output
         mock_cmd.return_value = _failure("pane not found")
         result = pane_output(pane=99)
         assert "Failed" in result
@@ -116,26 +116,26 @@ class TestPaneTools:
 
 
 class TestVoiceTools:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_notify.run_agentwire_cmd")
     def test_notify_with_target(self, mock_cmd):
-        from agentwire.mcp_server import notify_parent
+        from agentwire.mcp_notify import notify_parent
         mock_cmd.return_value = _success(delivered=True, target="main")
         result = notify_parent(text="hey", session="main")
         args = mock_cmd.call_args[0][0]
         assert args == ["notify-parent", "--to", "main", "hey"]
         assert "delivered" in result.lower()
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_notify.run_agentwire_cmd")
     def test_notify_without_target(self, mock_cmd):
-        from agentwire.mcp_server import notify_parent
+        from agentwire.mcp_notify import notify_parent
         mock_cmd.return_value = _success(delivered=True, target="parent")
         notify_parent(text="hey")
         args = mock_cmd.call_args[0][0]
         assert args == ["notify-parent", "hey"]
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_notify.run_agentwire_cmd")
     def test_notify_not_delivered(self, mock_cmd):
-        from agentwire.mcp_server import notify_parent
+        from agentwire.mcp_notify import notify_parent
         # safe_deliver refused (e.g. parked session) — surfaced, not hidden.
         mock_cmd.return_value = {"success": False, "delivered": False,
                                  "target": "main", "reason": "session is parked"}
@@ -150,16 +150,16 @@ class TestVoiceTools:
 
 
 class TestDesktopTools:
-    @patch("agentwire.mcp_server._portal_request")
+    @patch("agentwire.mcp_desktop._portal_request")
     def test_windows_list_empty(self, mock_req):
-        from agentwire.mcp_server import desktop_windows_list
+        from agentwire.mcp_desktop import desktop_windows_list
         mock_req.return_value = {"success": True, "windows": []}
         result = desktop_windows_list()
         assert "No windows" in result
 
-    @patch("agentwire.mcp_server._portal_request")
+    @patch("agentwire.mcp_desktop._portal_request")
     def test_open_session(self, mock_req):
-        from agentwire.mcp_server import desktop_open_session
+        from agentwire.mcp_desktop import desktop_open_session
         mock_req.return_value = {"success": True, "window_id": "win-1"}
         result = desktop_open_session(session="app", mode="monitor")
         mock_req.assert_called_once_with("POST", "/api/desktop/window/open", {
@@ -167,9 +167,9 @@ class TestDesktopTools:
         })
         assert "win-1" in result
 
-    @patch("agentwire.mcp_server._portal_request")
+    @patch("agentwire.mcp_desktop._portal_request")
     def test_write_artifact_success(self, mock_req):
-        from agentwire.mcp_server import desktop_write_artifact
+        from agentwire.mcp_desktop import desktop_write_artifact
         mock_req.side_effect = [
             {"success": True, "path": "/tmp/x.html", "url": "/artifacts/x.html"},
             {"success": True, "window_id": "art-1"},
@@ -178,16 +178,16 @@ class TestDesktopTools:
         assert "art-1" in result
         assert mock_req.call_count == 2
 
-    @patch("agentwire.mcp_server._portal_request")
+    @patch("agentwire.mcp_desktop._portal_request")
     def test_write_artifact_upload_failure(self, mock_req):
-        from agentwire.mcp_server import desktop_write_artifact
+        from agentwire.mcp_desktop import desktop_write_artifact
         mock_req.return_value = {"success": False, "error": "too large"}
         result = desktop_write_artifact(filename="x.html", html_content="data")
         assert "Failed" in result
 
-    @patch("agentwire.mcp_server._portal_request")
+    @patch("agentwire.mcp_desktop._portal_request")
     def test_close_window(self, mock_req):
-        from agentwire.mcp_server import desktop_close_window
+        from agentwire.mcp_desktop import desktop_close_window
         mock_req.return_value = {"success": True}
         result = desktop_close_window(window_id="win-1")
         assert "closed" in result
@@ -199,9 +199,9 @@ class TestDesktopTools:
 
 
 class TestSchedulerTools:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_status(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_status
+        from agentwire.mcp_scheduler import scheduler_status
         mock_cmd.return_value = _success(
             running=True, task_count=5, enabled_count=3,
             next_task="daily-check", next_in_seconds=120,
@@ -211,9 +211,9 @@ class TestSchedulerTools:
         assert "3/5" in result
         assert "daily-check" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_task.run_agentwire_cmd")
     def test_task_run_success(self, mock_cmd):
-        from agentwire.mcp_server import task_run
+        from agentwire.mcp_task import task_run
         mock_cmd.return_value = _success(
             status="complete", summary="All good", attempt=1,
         )
@@ -222,9 +222,9 @@ class TestSchedulerTools:
         args = mock_cmd.call_args[0][0]
         assert args == ["ensure", "-s", "app", "--task", "daily"]
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_task.run_agentwire_cmd")
     def test_task_run_exit_code_3(self, mock_cmd):
-        from agentwire.mcp_server import task_run
+        from agentwire.mcp_task import task_run
         mock_cmd.return_value = {"success": False, "error": "locked", "exit_code": 3}
         result = task_run(session="app", task="x")
         assert "locked" in result.lower()
@@ -236,9 +236,9 @@ class TestSchedulerTools:
 
 
 class TestHistoryTools:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_history.run_agentwire_cmd")
     def test_history_list(self, mock_cmd):
-        from agentwire.mcp_server import history_list
+        from agentwire.mcp_history import history_list
         mock_cmd.return_value = _success(items=[
             {"sessionId": "abc123", "firstMessage": "fix bug", "messageCount": 5},
         ])
@@ -246,9 +246,9 @@ class TestHistoryTools:
         assert "abc123" in result
         assert "fix bug" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_history.run_agentwire_cmd")
     def test_history_show(self, mock_cmd):
-        from agentwire.mcp_server import history_show
+        from agentwire.mcp_history import history_show
         mock_cmd.return_value = _success(
             sessionId="abc123", firstMessage="fix bug",
             gitBranch="main", messageCount=10,
@@ -264,9 +264,9 @@ class TestHistoryTools:
 
 
 class TestEmailTool:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_channels.run_agentwire_cmd")
     def test_email_send_full(self, mock_cmd):
-        from agentwire.mcp_server import email_send
+        from agentwire.mcp_channels import email_send
         mock_cmd.return_value = _success()
         result = email_send(body="hi", to="a@b.com", subject="test")
         args = mock_cmd.call_args[0][0]
@@ -274,17 +274,17 @@ class TestEmailTool:
         # Honest async boundary: accepted by provider, not "delivered" (#444).
         assert "accepted by provider" in result.lower()
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_channels.run_agentwire_cmd")
     def test_email_send_minimal(self, mock_cmd):
-        from agentwire.mcp_server import email_send
+        from agentwire.mcp_channels import email_send
         mock_cmd.return_value = _success()
         email_send(body="content only")
         args = mock_cmd.call_args[0][0]
         assert args == ["email", "--body", "content only"]
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_channels.run_agentwire_cmd")
     def test_email_send_with_attachments(self, mock_cmd):
-        from agentwire.mcp_server import email_send
+        from agentwire.mcp_channels import email_send
         mock_cmd.return_value = _success()
         email_send(body="see attached", attachments=["/tmp/a.pdf", "/tmp/b.csv"])
         args = mock_cmd.call_args[0][0]
@@ -293,17 +293,17 @@ class TestEmailTool:
         assert "/tmp/a.pdf" in args
         assert "/tmp/b.csv" in args
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_channels.run_agentwire_cmd")
     def test_email_send_plain_text(self, mock_cmd):
-        from agentwire.mcp_server import email_send
+        from agentwire.mcp_channels import email_send
         mock_cmd.return_value = _success()
         email_send(body="plain msg", plain_text=True)
         args = mock_cmd.call_args[0][0]
         assert "--plain" in args
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_channels.run_agentwire_cmd")
     def test_email_send_plain_text_false(self, mock_cmd):
-        from agentwire.mcp_server import email_send
+        from agentwire.mcp_channels import email_send
         mock_cmd.return_value = _success()
         email_send(body="html msg", plain_text=False)
         args = mock_cmd.call_args[0][0]
@@ -316,9 +316,9 @@ class TestEmailTool:
 
 
 class TestSchedulerEnableDisable:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_enable(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_enable
+        from agentwire.mcp_scheduler import scheduler_enable
         mock_cmd.return_value = _success()
         result = scheduler_enable(task="daily-check")
         mock_cmd.assert_called_once_with(
@@ -326,16 +326,16 @@ class TestSchedulerEnableDisable:
         )
         assert "enabled" in result.lower()
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_enable_failure(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_enable
+        from agentwire.mcp_scheduler import scheduler_enable
         mock_cmd.return_value = _failure("Task 'nope' not found")
         result = scheduler_enable(task="nope")
         assert "Failed" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_disable(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_disable
+        from agentwire.mcp_scheduler import scheduler_disable
         mock_cmd.return_value = _success()
         result = scheduler_disable(task="daily-check")
         mock_cmd.assert_called_once_with(
@@ -343,18 +343,18 @@ class TestSchedulerEnableDisable:
         )
         assert "disabled" in result.lower()
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_disable_failure(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_disable
+        from agentwire.mcp_scheduler import scheduler_disable
         mock_cmd.return_value = _failure("Task 'nope' not found")
         result = scheduler_disable(task="nope")
         assert "Failed" in result
 
 
 class TestSchedulerHistory:
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_history_success(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_history
+        from agentwire.mcp_scheduler import scheduler_history
         mock_cmd.return_value = _success(history=[
             {"task": "code-quality", "last_run": "2026-02-20T10:00:00",
              "last_status": "complete", "last_duration": 120, "run_count": 5},
@@ -367,23 +367,23 @@ class TestSchedulerHistory:
         assert "complete" in result
         mock_cmd.assert_called_once_with(["scheduler", "history", "--json"])
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_history_empty(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_history
+        from agentwire.mcp_scheduler import scheduler_history
         mock_cmd.return_value = _success(history=[])
         result = scheduler_history()
         assert "No run history" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_history_failure(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_history
+        from agentwire.mcp_scheduler import scheduler_history
         mock_cmd.return_value = _failure("board not found")
         result = scheduler_history()
         assert "Failed" in result
 
-    @patch("agentwire.mcp_server.run_agentwire_cmd")
+    @patch("agentwire.mcp_scheduler.run_agentwire_cmd")
     def test_scheduler_history_limit(self, mock_cmd):
-        from agentwire.mcp_server import scheduler_history
+        from agentwire.mcp_scheduler import scheduler_history
         mock_cmd.return_value = _success(history=[
             {"task": f"task-{i}", "last_run": f"2026-02-20T{10-i:02d}:00:00",
              "last_status": "complete", "last_duration": 60, "run_count": 1}
